Erasing qt (#245)
* enh(neb): neb objects updated to match engine refacto.
fix(neb): Contacts improved
enh(url_encode): New url encode function with associated tests.
enh(neb): get notification_number from engine #MON-4939

update heeders

* enh(variant): New variant class to replace QVariant

Conflicts:
	notification/src/loaders/command_loader.cc
	notification/src/loaders/dependency_loader.cc
	notification/src/loaders/macro_loader.cc
	notification/src/loaders/timeperiod_loader.cc

* enh(qt): A first step in the qt removing.

* enh(lua): No more qt object.

Conflicts:
	lua/test/lua.cc

* test(bbdo): New tests on the bbdo output class.

Conflicts:
	build/tests/CMakeLists.txt

* feat(bbdo): bbdo serialization is written without Qt

* enh(bbdo): New tests added.

* fix(bbdo): Checksum test fixed.

* fix(various): Compilation on CentOS7 fixed.

* fix(various): Compilation on CentOS7 fixed.

* WIP

Conflicts:
	sql/src/stream.cc
	storage/src/stream.cc

* test(processing/thread): New unit tests

Conflicts:
	build/tests/CMakeLists.txt

* enh(various): QThread and other Qt things removed

Conflicts:
	bam/inc/com/centreon/broker/bam/availability_thread.hh
	bam/inc/com/centreon/broker/bam/dimension_ba_timeperiod_relation.hh
	bam/inc/com/centreon/broker/bam/monitoring_stream.hh
	bam/inc/com/centreon/broker/bam/reporting_stream.hh
	bam/src/monitoring_stream.cc
	core/src/database.cc
	notification/src/stream.cc

* enh(core): QCoreApplication removed.

* enh(notification): Its compilation is disabled

Conflicts:
	bam/inc/com/centreon/broker/bam/availability_thread.hh

* enh(various): QObjects and friends removed.

* enh(config): type is an std::string

* enh(dumper): QDir removed to read directory content

Conflicts:
	bam/src/configuration/reader.cc
	bam/src/configuration/reader_v2.cc
	core/src/database.cc

* fix(dumper): Little headers cleanup

* enh(filesystem): new misc functions to work on filesystem

* test(misc): New tests on filesystem

Conflicts:
	build/tests/CMakeLists.txt

* enh(dumper): No more qt in dumper

* enh(graphite): No more Qt object in graphite connector

Conflicts:
	graphite/inc/com/centreon/broker/graphite/stream.hh

* enh(influxdb): No more qt objects

* enh(rrd): No more Qt objects

* enh(stats): No more qt object

* enh(tcp): No more qt objects

* fix(various): Various fix to compile

* enh(bam): almost all the qt objects are removed

Conflicts:
	bam/src/availability_builder.cc
	bam/src/configuration/reader.cc
	bam/src/configuration/reader_v2.cc

* enh(cmake): OFFLINE mode generalized on tests

* enh(simu): No more qt object

* fix(test): Various tests fixes.

* fix(cmake): Compilation improved with Asio

When Asio was downloaded, we could have some compilation that started
with the need of Asio whose download was not over.
A fix is also made in the processing::thread class.

* fix(variant): In some cases, we can have an integer for a boolean

* enh(cmake): QT_LIBRARY no more used in rrd/graphite/tcp

* test(neb): Unit tests on neb are restored.

* test(neb): New tests

On:
 * custom_variable_status
 * event_handler
 * flapping_status

* fix(cmake): json11 dependencies added.

* cleanup(core): Comments removed.

+++ neb/inc/com/centreon/broker/neb/statistics/generator.hh
@@ -53,8 +53,8 @@ namespace          neb {
       unsigned int interval() const throw ();
       void         interval(unsigned int value);
       void         remove(
-                     unsigned int host_id,
-                     unsigned int service_id);
+                     uint64_t host_id,
+                     uint64_t service_id);
       void         run();
       void         set(config::state const& config);
 
@@ -62,10 +62,8 @@ namespace          neb {
       unsigned int _interval;
       std::map<std::string, std::shared_ptr<plugin> >
                    _plugins;
-      std::map<
-             std::pair<unsigned int, unsigned int>,
-             std::shared_ptr<plugin> >
-                   _registers;
+      std::map<std::pair<uint64_t, uint64_t>, std::shared_ptr<plugin>>
+          _registers;
     };
   }
 }
